refactor: рефакторинг часть 4 - хедер, лайаут и стили
