--- src/components/app-header/app-header.module.css
@@ -1,50 +1,0 @@
-
-.header {
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-  background: #1c1c21;
-  width: 100%;
-}
-
-.panel {
-  padding: 16px 0;
-  width: 100%;
-  max-width: 1280px;
-  display: grid;
-  grid-template-columns: 1fr auto 1fr;
-  gap: 8px;
-  align-items: center;
-}
-
-.link {
-  padding: 16px 20px;
-}
-
-.navigation {
-  justify-self: start;
-}
-
-.logo {
-  justify-self: center;
-}
-
-.profile {
-  justify-self: end;
-}
-
-.flex_row {
-  display: flex;
-  flex-direction: row;
-}
-
-.non_list {
-  list-style: none;
-}
-
-.links {
-  composes: flex_row;
-  composes: non_list;
-  gap: 8px;
-}
-
