+++ src/components/app-layout/app-layout.module.css
@@ -1,0 +1,64 @@
+.header {
+  display: flex;
+  flex-direction: row;
+  justify-content: center;
+  background: #1c1c21;
+  width: 100%;
+}
+
+.panel {
+  padding: 16px 0;
+  width: 100%;
+  max-width: 1280px;
+}
+
+.links {
+  list-style: none;
+  display: grid;
+  grid-template-columns: fit-content(167px) fit-content(167px) auto minmax(0, 167px) fit-content(167px);
+  gap: 8px;
+  align-items: center;
+  grid-template-areas:
+    "linkHome linkFeed logo . linkProfile";
+}
+
+.link {
+  padding: 16px 20px;
+  width: fit-content;
+  display: flex;
+  flex-direction: row;
+}
+
+.link_home {
+  composes: link;
+  grid-area: linkHome;
+}
+
+.link_feed {
+  composes: link;
+  grid-area: linkFeed;
+}
+
+.link_logo {
+  justify-self: center;
+  grid-area: logo;
+}
+
+.link_profile {
+  composes: link;
+  justify-self: end;
+  margin-left: auto;
+  grid-area: linkProfile;
+}
+
+.main {
+  margin: 0;
+  padding: 40px 0 0;
+  width: 100%;
+  max-width: 1280px;
+  box-sizing: border-box;
+  height: 100%;
+  max-height: calc(100% - (52px + 88px));
+  display: flex;
+  flex-direction: row;
+}
