+++ src/components/navigation-link/navigation-link.module.css
@@ -4,3 +4,7 @@
   flex-direction: row;
   color: inherit;
 }
+
+.label {
+  white-space: nowrap;
+}
